+++ DashAI/back/main.py
@@ -4,13 +4,32 @@ from fastapi import FastAPI
 from fastapi.responses import RedirectResponse
 from starlette.responses import FileResponse
 
+from DashAI.back.models import SVC, KNeighborsClassifier, RandomForestClassifier
+from DashAI.back.registries import ModelRegistry, TaskRegistry
 from DashAI.back.routers import datasets, experiments
+from DashAI.back.tasks import (
+    TabularClassificationTask,
+    TextClassificationTask,
+    TranslationTask,
+)
 
-# TODO These imports should be removed because they are unused, but currently needed.
-from DashAI.back.tasks.tabular_classification_task import TabularClassificationTask
-from DashAI.back.tasks.task import Task
-from DashAI.back.tasks.text_classification_task import TextClassificationTask
-from DashAI.back.tasks.translation_task import TranslationTask
+task_registry = TaskRegistry(
+    initial_components=[
+        TabularClassificationTask,
+        TextClassificationTask,
+        TranslationTask,
+    ],
+)
+
+model_registry = ModelRegistry(
+    task_registry=task_registry,
+    initial_components=[
+        SVC,
+        KNeighborsClassifier,
+        RandomForestClassifier,
+    ],
+)
+
 
 app = FastAPI(title="DashAI")
 api = FastAPI(title="DashAI API")