+++ DashAI/back/models/scikit_learn/svc.py
@@ -1,20 +1,18 @@
 import json
 
-from sklearn.svm import SVC
+from sklearn.svm import SVC as _SVC
 
-from DashAI.back.models.classes.sklearn_like_model import SklearnLikeModel
-from DashAI.back.models.classes.tabular_classification_model import (
-    TabularClassificationModel,
-)
+from DashAI.back.models.scikit_learn.sklearn_like_model import SklearnLikeModel
+from DashAI.back.models.tabular_classification_model import TabularClassificationModel
 
 
-class SVM(SklearnLikeModel, TabularClassificationModel, SVC):
+class SVC(TabularClassificationModel, SklearnLikeModel, _SVC):
     """
     Support vector machine. Supervised learning algorithm that separates
     two classes in two spaces by means of a hyperplane. This hyperplane is
     defined as a vector called support vector.
     """
 
-    MODEL = "svm"
+    MODEL = "SVC"
     with open(f"DashAI/back/models/parameters/models_schemas/{MODEL}.json") as f:
         SCHEMA = json.load(f)