+++ DashAI/back/models/text_classification_model.py
@@ -2,3 +2,5 @@ class TextClassificationModel:
     """
     Class for models associated to TextClassificationTask
     """
+
+    _compatible_tasks = ["TextClassificationTask"]