+++ DashAI/back/tasks/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+from DashAI.back.tasks.base_task import BaseTask
+from DashAI.back.tasks.tabular_classification_task import TabularClassificationTask
+from DashAI.back.tasks.text_classification_task import TextClassificationTask
+from DashAI.back.tasks.translation_task import TranslationTask